added proper example
