+++ example/go.mod
@@ -1,0 +1,8 @@
+module github.com/Kansuler/octobe/example
+
+go 1.21.5
+
+require (
+	github.com/Kansuler/octobe/v2 v2.0.0-20231222135250-80258178b429 // indirect
+	github.com/go-chi/chi/v5 v5.0.11 // indirect
+)
